+++ CHANGELOG.md
@@ -31,6 +31,7 @@ The official release will have a defined and more stable API. If you are already
   * UI: Added feedback after killing a step
   * Killing a `shell/bash`-step now kills the whole process tree spawned by it. This helps in cases where the step spawns
     longer-running processes and bash doesn't pass on the TERM signal to its children)
+  * Fixed UI bug where pipeline was no longer visible for long step output
 
 ## 0.5.6