+++ CHANGELOG.md
@@ -21,6 +21,7 @@ The official release will have a defined and more stable API. If you are already
   * Improved calculation of build duration for retriggered pipelines (#30)
   * UI: collapsing child steps by default (#59)
   * Add feature to alias build steps in UI:
+
     ```clojure
     ; this displays "trigger" instead of "either" in UI
     (alias "trigger"
@@ -28,6 +29,7 @@ The official release will have a defined and more stable API. If you are already
         wait-for-manual-trigger
         wait-for-repo))
     ```
+  
   * UI: Added feedback after killing a step
   * Killing a `shell/bash`-step now kills the whole process tree spawned by it. This helps in cases where the step spawns
     longer-running processes and bash doesn't pass on the TERM signal to its children)